refactor: 💡 separated out the implementation in componnets
